Calculate tables until 2037.
Currently, we don't have a way to calcuate timezone offsets going forward in
time. We're supposed to use the most recent rules and use them to project the
timezone offset and daylight savings time adjustments into the future, even
though the rule are subject to change.

Suddendly, however, I have a few early adopters, so calculating a table that
reaches out to the 32-bit overflow of seconds since the epoch, the [2038
bug](http://en.wikipedia.org/wiki/Year_2038_problem), is a stop gap to create a
work library that can work with contemporary or historic dates.

    File                             Raw         Min         Zip     Min/Zip
    lib/transitions.js        6791/06.63  3044/02.97  2154/02.10  1312/01.28