Converting to JavaScript.
I'd like to have control over the generated JavaScript, because I'd like to make
it as small as possible.

My experience with minifying JavaScript so far has shown me that the only thing
that makes it smaller is removing logic. `gzip` is so efficient that it doesn't
matter so much if you're verbose, using long names, or fiddly constructs. You
wouldn't save four bytes everywhere if the `function` keyword was replaced with
`func`.

Converting from objects as JSON to arrays, for example, doesn't give you as much
savings as you'd imagine after `gzip`, because `gzip` is going to eliminate the
heft of the object key names, so long as they are redundant.

Thus, I'm pretty sure that I could have a small program with CoffeeScript. If
CoffeeScript is more verbose at times, using boilerplate for testing against
both `undefined` and `null`, `gzip` will smash it up good. It looks heavy, but
it is a common construct, repeated frequently, and it gets squished.

If I'm still interested in demonstrating the viability of CoffeeScript later on,
it is easier to convert from JavaScript to CoffeeScript. Once I have a tiny
JavaScript library, I can come back and see how much it grows when converted to
CoffeeScript.

In the mean time, I really want to see what I'm doing, so I can learn what sort
of JavaScript makes for micro JavaScript.

I'm going to start by adding my CoffeeScript generated JavaScript. This commit
will add the JavaScript, the next commit will de-dent it, then I'll start to
convert the code a function at a time, into hand-crafted JavaScript.

Here's where I'm at now with file size.

    File                             Raw         Min         Zip     Min/Zip
    lib/timezone.js          33014/32.24 12395/12.10  7304/07.13  4921/04.81