Fixed zone table scope.
Including a timezone rule set table would make that rule set table visible to
any instance of the `tz` function, not just a newly created partial function.
This was because I was not copying the new timezone rule set table into the
request, which is the context of a new partial definition. I was copying it into
the context, which would expose it to the `tz` function used to create the
partial function.

    File                             Raw         Min         Zip     Min/Zip
    lib/timezone.js          28244/27.58 11984/11.70  6927/06.76  4780/04.67