Tidy of `convert`.
General tidy of `convert`, mostly to remove CoffeeScript array iteration so I
can see what's going on.

Changed one test of `indexOf` to use bitwise negation of the result, which is
the style I prefer.

The `adjustments` array does not need to be part of the `request` object. It can
be a local variable and as such, it can get minified.

    File                             Raw         Min         Zip     Min/Zip
    lib/timezone.js          28066/27.41 11911/11.63  6931/06.77  4733/04.62