Remove closure around adjust, tidy helpers.
Changed the function declaration style of the `append` and `extend` helpers.
Tidies the loop in `extend`, removing unnecessary local variables.

Removed the closure from around adjust exposing it and all its helpers in the
scope of the module definition closure. This is an initial step before further
rewriting, so that significant changes are not lost the diffs due to in
whitespace changes.

    File                             Raw         Min         Zip     Min/Zip
    lib/timezone.js          27973/27.32 11831/11.55  6925/06.76  4732/04.62