Smaller implementation of `adjustment`.
Couple of modifications to make date math require less code.

Simplified the date math language. Not as many synonyms. Not as many literals in
the code.

Building the regular expression from a string. Using the string that defines the
alternation that matches the time units to create a lookup table to determine
which unit was matched by the regular expression. No need to express the
language keywords in two different forms.

Using `Date` to do year, month, and date math, instead of incrementing and
carrying an array of integer values for year, month, date. Less logic means less
code which means less size.

Using `parseInt` to determine the sign matched by the regular expression instead
of a lookup object.

Eliminated variables, inlining some arithmetic. Less variables is less size.

    File                             Raw         Min         Zip     Min/Zip
    lib/timezone.js          24955/24.37 11158/10.90  6371/06.22  4479/04.37