Remove `actualize` for now. Tidy timezone lookup.
Tidy timezone lookup somewhat. The CoffeeScript generated JavaScript was already
pretty well formatted.

Removing `actualize` for now. It is an old implementation of the function that
is not currently used. When we do get back around to actualizing the tables as
needed, we're going to use a newer, smaller implementation from
`lib/transitions.js`.

    File                             Raw         Min         Zip     Min/Zip
    lib/timezone.js          22340/21.82 10070/09.83  5928/05.79  4065/03.97