Limit date parsing to RFC 3999.
Fortunately, there is such a thing as RFC 3999 that defines the reasonable
subset of ISO 8601 that we tend to use. This way, we don't have to say that we
parse a subset ISO 8601, we can say instead that we parse the entirety of RFC
3999. The RFC 3999 explains the reasons for using a subset of ISO 8601 with
authority. We are not saddle with the burden of explanation.

Nor do we have to bloat the library with absurd date formats that no one will
ever use, so that we can say that we parse 8601 without caveats. No one will
ever want to parse a date with week number and day of week, but if we don't
parse it, we can't make a succinct statement about our date string formats
relationship to this standard.

I'd not actually read RFC 3999. I knew that it supported using a space in place
of the `T` to make the date slightly more legible. I didn't realize it defined a
subset. If someone had asked me, I might have said that it was merely an
extension that permitted the space.

I'm happy that it is a well reasoned subset. The discussion in RFC 3999 is as
good a discussion of the experiences of storing time as strings as any I've
encountered so far. It will be cited in the landing page for this library.

I doubt that I'll bother with ISO 8601 myself. If someone wants to write an ISO
8601 parser, they can do date math to navigate by days of the year and weeks.

Also, ISO 8601 is only half about date formats. The rest of it is about
durations and intervals, which is beyond the scope of my interest in computer
storage of dates and times for this lifetime.

The RFC 3999 date parser correctly parses timezone offsets, so it closes #8.

The RFC 3999 date parser correctly parses a full date accounting for the
timezone, converting from wall-clock time to POSIX time. No conversion according
to timezone is performed if a timezone offset is specified explicitly in the
date string, either by timezone offset or by using the letter `Z` to indicate
`UTC`.

The timezone aware parsing us closer to closing issue #3, but we still need to
test the determination of an implicit date.

No longer parsing RFC 822. Moved the code off to the side. Will either make it a
pluggable parser, or a parser that parses RFC 822 and turns it into RFC 3999
for use by `tz`.

Better still, an RFC 822 parser can emit the intermediate array format that the
RFC 3999 parser internal to `tz` emits.

    File                             Raw         Min         Zip     Min/Zip
    lib/timezone.js          18934/18.49  8762/08.56  5173/05.05  3599/03.51