Tidy of `format`.
Attaching the padding size to the date formatting function as an object property
so I don't have to keep a separate object to use as a map of format specifiers
to padding sizes. Attaching the style and padding properties to the formatting
function will make it easier to split a set of the formatting specifiers out
into an extension.

The regular expression that matches format strings is now built from a string
using format specifiers gathered the specifier map. This regular expression is
rebuilt when new format specifiers are added. One more step to splitting up
formatting into basic and extended.

We're going to split foramtting into basic and extended. Basic foramtting is
anything numeric that does not require a mess of calculation. Extended
formatting will include the day of year, week of year, and localized format
specifiers.

The day of year and week of year calculations are not a common use case, and the
week of year functions are rather hefty at the moment.

Localization means asynchronously loading locale JSON, which is by its nature,
something extra. There are use cases for this library where it is doing
calculations without localization strings, such as logging, where the basic
formatting can still create human readable strings, provided that human is a
system administrator.

    File                             Raw         Min         Zip     Min/Zip
    lib/timezone.js          19223/18.77  8927/08.72  5204/05.08  3613/03.53