Break full GNU `date` out into `date.js`.
Locale support is hefty. If you want a little you get a lot, so if you really
don't want a lot, you're going to not want any. When you do want locale support,
you can have it, along with week of year, and day of year, with a single
additional library.

The `date.js` library is a formatting extension library for the `tz` function
that implements all of the rest of GNU `date`.

Made `weekOfYear` a little smaller by removing some unnecessary local variables.

Also, I found found a bug I'd left in my tests. Math across daylight savings
time clock transitions was displaying the correct timezone offset, but the wrong
timezone acronym. At the time I wrote the test, I was about to reimplement
timezones lookup, so I made the broken output the fixture. Left a `TODO` and
everything.

Now it's fixed.

    File                             Raw         Min         Zip     Min/Zip
    lib/timezone.js          11988/11.71  5500/05.37  3464/03.38  2415/02.36
    date.js                   4591/04.48  3077/03.00  1256/01.23  1034/01.01