Fix array date creation and test it.
Creating a date from an array of integer field values, where the integers are
integers or their string representations. This is the same logic used internally
to build a milliseconds since the epoch offset from a parsed RFC 3999 date
string, so the `tz` function can consume the output of other date parsers that
create this array of integers format.

The array of integers format includes a capacity for zone offsets, treating the
three digits after either a `'-'` or `'+'` character as the hours, minutes and
seconds of a timezone offset. The sign character is used to determine the sign
of the offset.

If there are fewer than five numbers before a sign, then it is assumed that the
date portion of the date was not specified, only the time. If this is case, then
the year, month, and date of the applicable timezone according to the applicable
clock is used to complete the date.

This method of specifying a date is for use with other date parsers, or UI
controls like date pickers. It doesn't second guess its input in any way.

This was all implemented when back when ISO 8601 was eliminated and the RFC 3999
only parser was introduced. It's now exposed to the `tz` caller. Well, it was
already exposed to the `tz` caller, except that it totally didn't work.

Progress on #7. Need to consider application of timezone.

    File                             Raw         Min         Zip     Min/Zip
    lib/timezone.js          12029/11.75  5506/05.38  3458/03.38  2411/02.35
    date.js                   4591/04.48  3077/03.00  1256/01.23  1034/01.01