Smaller zone offset format function.
The `splitOffset` function was used only by the format function for the `%z`
format specifier, so its logic was placed into the format function. The logic to
split the offset into hours, minutes an seconds was rolled into a loop.

The formatting logic itself uses the length of the `delimiters` string, the
colons between the `%` and `z` to determine where to slice the array. The double
regular expressions are used to strip the zone offset in the case of the
just-make-it-long-enough option.

Not sure if I need `UTC` anymore, so I got rid if it.

    File                             Raw         Min         Zip     Min/Zip
    lib/timezone.js          11200/10.94  5114/04.99  3288/03.21  2283/02.23
    date.js                   4591/04.48  3077/03.00  1256/01.23  1034/01.01