Further tidy of `format`.
Padding is not down using a negative parameter to `slice`. No need to build
padding strings. We know that we'll only ever need eight spaces, or eight zeros.
The `splice` method minzips better than the padding functions.

Got rid of the lookup objects for transforms and padding functions. Expressing
this logic using conditionals instead.

Replaced an if/else with a ternary operator in `convertToPOSIX`.

    File                             Raw         Min         Zip     Min/Zip
    lib/timezone.js          10444/10.20  4922/04.81  3105/03.03  2208/02.16
    date.js                   4591/04.48  3077/03.00  1256/01.23  1034/01.01