Argument flattening, smaller `zoneinfo`.
Fixed argument flattening. Was testing loading a world of zone info to discover
that I'd transposed some calls in a method chain.

Tidy `zoneinfo` for some size savings. Using single literals for time instead of
a string of multiplication of literals. Seeing the multiplication was my way of
seeing the using of time; sixty times a thousand milliseconds is a minute. I
know the products by heart now, though. They're no longer magic numbers to me.

    File                             Raw         Min         Zip     Min/Zip
    lib/timezone.js          10052/09.82  4631/04.52  3040/02.97  2123/02.07
    date.js                   4472/04.37  3058/02.99  1202/01.17  1012/00.99