Smaller week of year formatting functions.
Revisited the logic for ISO week. Found some savings by eliminating local
variables, reordering the statements.

I wanted to consider the [other means of determining the if the last ISO week of
the year is a leap week](http://en.wikipedia.org/wiki/ISO_week_date), to see if
any of them would take less code. Nothing occurred to me. Testing the first day
of year against the current year, or the previous year, repeats the same test,
which compresses better an alternative that would check the first and last days
of the year.

Thought for a long time about how simplify the ranged comparison to determine
the first ISO week. Confused as to why the lower half of the test is greater
than Monday, instead of greater than or equal to Monday. It's because
`weekOfYear` returns `1` if New Years Day is Monday, which is already the
correct offset for determining the ISO week of year.

Closes #20.

    File                             Raw         Min         Zip     Min/Zip
    lib/timezone.js          10052/09.82  4631/04.52  3040/02.97  2123/02.07
    date.js                   4302/04.20  2949/02.88  1150/01.12   969/00.95