Support for AMD.
Added support for AMD to `date.js` and it's still under 1k. Added AMD to
`lib/timezone.js`.

AMD is new to me. If it gains adoption, it may solve the problem of how to
provision zone rule tables. It won't be on me to provide an asynchronous loader.

I'm sure there are a lot of opinions about AMD out there. I'm not curious to
know what they are at the moment. It has the appearances of a standard, plus it
shows up in a handful of the projects featured on
[microjs](http://microjs.com/). There are alternatives. Until someone says
otherwise, this alternative is as good as any.

I don't want to defer this decision. I don't want to punt to the application
developer, have them choose the module system that suits them. I want to
honestly account for the cost of module loading in the braggable minzip sizes.

If the application developer has their solution of their for module loading,
they can make the conversion on their own. I'll support the two favors of
RequireJS and direct inclusion via `script` tags.

To the application developer, I say, you can use AMD to load rule set tables. If
you don't use AMD, then use the asynchronous mechanism you've chosen for your
application. For some applications, Timezone will require you to load additional
resources as needed. This is the nature of web applications. I hope you came
prepared.

    File                             Raw         Min         Zip     Min/Zip
    lib/timezone.js          10211/09.97  4731/04.62  3083/03.01  2149/02.10
    date.js                   4465/04.36  3061/02.99  1194/01.17  1004/00.98