Implement parse time without year.
Instead of having `makeDate` determine if the year is missing, we leave that
determination up to the date parser. Not all date string formats are going to
support partial representations. RFC 822, for example.

Working from my assumptions, based on my knowledge of ISO 8601, I'd implemented
an RFC 3999 date parser that would parse time without a date. Upon closer
reading of RFC 3999, I can see that this is not specified in RFC 3999. It states
that partial representations run counter to its goals of simplicity.

I've left partial date parsing in for now, using the logic from `makeDate`, but
it will remain for a limited time, for posterity. It is not a useful behavior.
It is simple enough for the application developer to fill in partial date.

    var vilnius = tz("Europe/Vilnius");
    var noonInVilnius = vilnius(vilnius("%Y-%m-%d 12:00:00"));

I moved the timezone application from the RFC 3999 parser, into `makeDate`. If
the date array includes a `+` or `-` string, it will apply a timezone offset
specified within the date array. If not `makeDate` will adjust the date using
the effective timezone.

The `makeDate` function is exposed to the application developer. They can pass
in an array of the `makeDate` format as the date argument. You can write a date
parser that returns a date array, and feed that date array to `tz` to get
timezone adjusted POSIX time.

Closes #3.
Closes #7.

    File                             Raw         Min         Zip     Min/Zip
    lib/timezone.js          10251/10.01  4723/04.61  3068/03.00  2154/02.10
    date.js                   4465/04.36  3061/02.99  1194/01.17  1004/00.98