GNU date complete, canned formats.
Created two custom format specifiers using non-alphabetic characters.

The `%.` format specifier prints zero-padded milliseconds. Milliseconds are the
prevision supported by JavaScript `Date`, but GNU `date` does not have way to
specify milliseconds, only nanoseconds.

The `%$` specifier prints the zone offset in zero padded hours and minutes,
colon delimited, unless the zone offset is zero, in which case it prints a `Z`.
This allows us to create RFC 3999 dates.

The `%:$` specifier prints the zone offset in zero padded hours and minutes,
colon delimited, unless the zone offset is zero, in which case it prints a `Z`.
If there are seconds in the zone offset, then it adds the seconds zero padded,
delimited by a colon. This allows us to create RFC 3999 dates with our extension
for down to the second timezone offsets.

Not feeling that canned formats for RFC 3999 are necessary. RFC 3999 can be
assembled with minimal chatter using some of the default specifiers built into
GNU `date`. RFC 3999 according to the letter of the RFC is `"%FT%T%$"`, or
optionally `"%F %T%$"`. The application developer that needs more precision from
a date strings can added milliseconds and timzeone offsets with seconds with
`"%FT%T.%.%:$"`.

It's awfully easy to configure a preferred RFC 3999 timestamp with a constant,
or a partial `tz` function. I'd rather not extend GNU `date` too much. I'm not
in a position at the moment to say which is the best timestamp format, so in the
end, we'd probably pluck one character from top row of the keyboard, and assign
it to each permutation.

I added the format specifiers for tabs and newlines.

I had `%r` wrong. It's not static. It's set by the locale. Added a 12 hour clock
format to `en_US` and `pl_PL`.

Not going to do `#` flag. It twiddles the case, mostly to upper, sometimes to
lower. Someone will have to miss it. Seems fiddly and useless. It's locale
specific, so it would bulk up the locales. It's stated as optional. What good
does it do to do the opposite of what's expected? Might ad it to the format
pattern matcher, but then ignore it.

The `E` and `O` modifiers are well out of scope. Not going to offer alternative
locale formats. If you want an alternative locale format, then provide an
alternative locale definition.

Not going to do GNU width padding for now, but that actually seems useful.

This completes the audit. I have looked over the GNU `date` documentation and
made a decisions about each date format.

Implemented a generic regular expression that does not assert that there is a
format specifier function using alternation within the regular expression. If no
format specifier function is mapped to the character after a `%`, the swallowed
`%` is added to the output string along with the unmapped character.

Closes #9.
Closes #19.

    File                             Raw         Min         Zip     Min/Zip
    lib/timezone.js          10579/10.33  4819/04.71  3123/03.05  2173/02.12
    date.js                   4515/04.41  3090/03.02  1200/01.17  1007/00.98