Olson transition test now in Ruby for Travis CI.
Reimplemented the test that verifies the clock transition we've discovered in
the Olson files in Ruby. Travis CI doesn't like our bash program. I suspect that
it doesn't like programs that fork a bunch, and our bash test forks `date` 48266
times currently.

Tried to write the test in Perl, but Perl's `strftime` does not support `%::z`.

Ruby's does, but only in 1.9.3. The next few commits are going to explore using
`rvm` on the Node.js worker at Travis CI to install Ruby 1.9.3, for the sake of
this one test.