Install Ruby 1.9.3 using `apt-get` at Travis CI.
It's not enough to use the `rvm` YAML property. It's not recognized on a Node.js
worker. Wasn't able to use `rvm`, because there wasn't a ruby 1.9.3 already
installed in the `~/.rvm` directory of Travis CI's `vagrant` user. Couldn't
install one, because building Ruby 1.9.3 takes too long.

Installing a 3rd party Ruby 1.9.3 from
[Brightbox](http://blog.brightbox.co.uk/) using their Ubuntu Personal Package
Archive. Installing Ruby 1.9.3 using apt-get gets Ruby 1.9.3 installed fast
enough for a Travis CI test run.

We need Ruby 1.9.3 for the `%::z` format specifier. The `strftime` in 1.9.2 and
earlier does not support it.

I've created a bash program `bin/test` that will perform this apt-get install if
we're on Travis CI.

The clock transition tests are key tests of the integrity of this library. It
was well worth the effort to add those tests to continuous integration.