Heuristics assertions.
Created a program to report the smallest interval between meaningful changes in
timezone offset or timezone abbreviation. The shortest interval is 6.125 days in
`Asia/Yerevan`. This confirms that we can sort transitions by an actualized
date, and that if we're off by some combination of timezone offset and daylight
savings time, we don't muddle the order of events.

When we order rules, we're not always able to calculate the exact time for each
of our clocks, wall-clock, standard and POSIX, because it depends on the rules
in effect, that is, the timezone offset plus any daylight savings. We need to
know what rules came before the current rule. To do this we sort the rules, but
we don't know exact times, because we're unsure of offset and savings. Thus, we
sort by date, instead of by date and time.

These tests show that we can indeed sort by date and obtain the order in which
daylight savings time rules are applied.

Closes #16.