Test millisecond before transition, fix MIN_VALUE.
`Number.MIN_VALUE` is the [smallest positive numeric value representable by
JavaScript](https://developer.mozilla.org/en/JavaScript/Reference/Global_Objects/Number/MAX_VALUE)
, not the smallest negative value. The smallest negative value is
`-Number.MAX_VALUE`.

Added an initial record to the skip list for the timezone abbreviation before
the adoption of standard time.

The find function can now find the correct offset and savings for the
millisecond before a clock transition.

The find function was failing for dates that fell on the new year. For
`America/Detroit`, the shift from `Detroit` rules to `US` rules at New Year's
1975 cased the `applicable` function to skip the fall back date in October 1974.

Rather than continue to fiddle with the date math to get the correct current
year for wall-clock specified year ranges for a POSIX time, I decided to use a
range of three years, including the two years surrounding the year indicated by
the POSIX time. It creates a larger a set of applicable intervals which we may
seek to prune in time, but for now it gets us to a working algorithm that we can
optimize later with tests in place, and all searches correctly implemented.