Now correclty finding wallclock time.
Exclude rules that are younger than the start of the current zone entry record
according to the clock used to specify the rule.

Surprised this is not breaking because standard time is not available in the
zone entry record. Dumb luck. We're going to try to translate standard time to
rules to wall-clock time, so hopefully this will not be an issue.