Skip list in `lib/timezone.js`.
The skip list method of determining zone offset and abbreviation has now
replaced the binary search in `lib/timezone.js`. `lib/timezone.js` is now much
fatter than it has been, with a bloaty implementation of everything involved.

The new zone offset and abbreviation lookup requires that the abbreviations are
generated from the formats. This logic will always be larger than the clock
transition table's implementation, since the formatted abbreviations were part
of the lookup table.

The offset in the entry record no longer includes the daylight savings.

Created `bin/zones.js` to export the skip list to the `zones` directory,
replacing our sorted tables of clock transitions with the skip lists in the zone
files for each locality. The function `lib/transitions.js` now returns an object
containing both the clock transitions table and the skip list. The clock
transitions table is still used for assertions about our heuristics, and to
generate our file containing all the clock transitions that we use to test our
parser against GNU `date`.

A squish of all this code will come in a separate working branch, possibly after
another release, certainly after adding the GNU extension for user specified
padding extensions. Allow the code to expand a little, while you assert that the
skip list is working and that the implementation of `strftime` is complete.

    File                             Raw         Min         Zip     Min/Zip
    lib/timezone.js          13328/13.02  6243/06.10  3929/03.84  2721/02.66
    lib/transitions.js       12769/12.47  5799/05.66  3381/03.30  2207/02.16
    date.js                   4515/04.41  3090/03.02  1200/01.17  1007/00.98
    timezones/asia.js      156900/153.22 76694/74.90  8713/08.51  7343/07.17
    zones/America/Detroit.js  2994/02.92  2483/02.42   560/00.55   546/00.53

The size of `lib/transitions.js` doesn't matter anymore. We're never going to
expose this to an application now. It is only a utility to generate the files.
Putting in the report one last time as a bookend.

Closes #21.