Export zones into Olson as JSON.
Pausing for a minute to consider where we're going with this. For Node.js, a
link can simply be a file that requires the actual file. In the browser, the
link will have to duplicate the data, so that there is not second trip to fetch
it. The delay of the second request is more expensive than the bytes of storage.

The storage is a trivial cost in Node.js, of course. Their may be some advantage
to indicating to the application that the zone rule set name is an alias. On my
Fedora machine, links are represented as hard links to the actual zone rule set,
so there's no why to which name was paired with the actual rule set in the Olson
files, and which one was an alias.