Creating linked zone files.
Linked zone files are the existing zone files with the alias name, and with a
zone entry who's key name is the alias name.

We're not doing any sort of actual linking here. The zoneinfo files on my Fedora
16 install are hard linked, but we're not going to have hard links on all
platforms, plus our current data structure requires that the zone name is in the
file.

Closes #11.