User specified padding GNU extension.
Implemented the user specified padding GNU extension. I added it because through
it, you can get a GNU `date` compatible format for milliseconds with `%3N`. With
nanoseconds, if the specifier is shorter, the nanoseconds are truncated. I'm
finding that this behavior is specific to nanoseconds. You can't truncate full
years, `%2Y` does not print `12` for the time of this commit.

Removed our `%.` specifier for milliseconds. We will use the GNU `date`
nanoseconds specifier with a padding of three for milliseconds instead.

Also, rewrote the format function to use `String.replace`. Removed our previous
padding strategy that prepended fill characters and applied `String.slice` with
a negative index from the `format` function. Left that strategy in place in the
`z` format specifier function.

Closes #22.

    File                             Raw         Min         Zip     Min/Zip
    lib/timezone.js          13081/12.77  6064/05.92  3841/03.75  2646/02.58
    date.js                   4515/04.41  3090/03.02  1200/01.17  1007/00.98
    timezones/asia.js      156955/153.28 76734/74.94  8746/08.54  7372/07.20
    zones/America/Detroit.js  2994/02.92  2483/02.42   560/00.55   546/00.53