Synopsis rough draft.
Got a feel for what I'm trying to say in the synopsis, and how I'm going to say
it. I'm about to delete a bunch of repetitive stuff, so I'm going to commit what
I have.

Able to move away from discussing motivations. Trying to imagine myself writing
for people who've already adopted the library, or rather, who've already adopted
the practices espoused by the library, and are looking to see how it use it.