Second draft of synopsis.
Definitely in the order I need it to be in, with all the unused stuff
gravitating toward the bottom the of document.

This document has a lot of examples that are US examples. I'd like to use more
examples from other parts of the world, once I get the synopsis working. I can
imagine that people looking for localized dates will be reassured to find
examples using locales and timezones outside of a single country.